requirements.txt
mss
opencv-python
screeninfo
pillow
tkthread
keyboard